ml_api/requirements.txt
open_clip_torch==2.31.0
pillow==11.1.0
fastapi==0.115.8
uvicorn==0.34.0
pydantic==2.10.6
pydantic-settings==2.7.1
python-multipart==0.0.20
aiofiles==24.1.0
numpy==2.2.3
requests